requirements.txt
# Core dependencies - compatible with Python 3.13
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
networkx>=3.2

# Deep learning (optional - will skip if installation fails)
torch>=2.1.0
torch-geometric>=2.4.0

# Visualization
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Explainability
shap>=0.43.0
lime>=0.2.0.1

# Data processing
scipy>=1.11.0
joblib>=1.3.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0